requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
langchain-google-genai
langchain-community
langchain-text-splitters
pymupdf
supabase
Pillow
gunicorn